requirements.txt
cirq-core
openfermion
git+https://github.com/IntelLabs/mat2qubit
